fix: harden PlayerCore state machine, cancellation and teardown
Engine deep-review remediation (VideoPlayerCore):
- H4/H5/H7: introduce prepareGeneration token in start(). Restore .idle
  on cancellation, and suppress stale cleanup / failure surfacing when a
  newer start has superseded the in-flight one.
- H3: add PlayerError.classify (NSURLError -> network/auth, AVFoundation
  -> decoding) and route mapToPlayerError through it so callers can
  distinguish failure causes instead of flattening to .unknown.
- H10/M1: make dispose() async and call engine.stop(.appLifecycle)
  idempotently before tearing down, so abnormal teardown no longer leaves
  a live player session / proxy timer.
- M3: preserve terminal (.finished/.failed) status against late
  bufferingDidChange events.
- M4: keep a resolved duration when an unconfirmed duration(0) arrives.
- M10: PlayerFeaturePolicy.maxPlaybackRate Float -> Double to remove the
  binary float boundary error on policy-rate comparison.

+++ Sources/VideoPlayerCore/Domain/PlayerFeaturePolicy.swift
@@ -10,7 +10,9 @@ import Foundation
 
 public struct PlayerFeaturePolicy: Equatable, Sendable {
     public let allowsBackgroundPlayback: Bool
-    public let maxPlaybackRate: Float
+    /// M10 — `Double`로 통일. 과거 `Float`이라 `Double(Float(1.4))`의 이진 부동소수 오차로
+    /// 정책 상한 경계(1.4·1.6 등) 비교가 어긋날 수 있었다.
+    public let maxPlaybackRate: Double
     public let allowsAutoplay: Bool
     public let skipInterval: TimeInterval
     public let nextEpisodeButtonLeadTime: TimeInterval
@@ -25,7 +27,7 @@ public struct PlayerFeaturePolicy: Equatable, Sendable {
 
     public init(
         allowsBackgroundPlayback: Bool,
-        maxPlaybackRate: Float,
+        maxPlaybackRate: Double,
         allowsAutoplay: Bool,
         skipInterval: TimeInterval = 10,
         nextEpisodeButtonLeadTime: TimeInterval = 30
